Created the `launch_nextflow()` function.
The `launch_nextflow()` function works as it should. The one issue I see
going into the future is not being able to run the scripts/ workflow
directly from GitHub. This should be okat as long as I remember to make
the (hopefully soon working) Nextflow docker image download the correct
scipt from the s3 bucket.