Minor Changes
Just made some small changes to the parameters, and fixed the annotation
portion of the workflow. Apparently it was annotation the concatinated
vaf and just naming it as the ExpansionHunter vcf... whoops. I will need
to build a Python launcher for this workflow as well. Once that is
started I will add the `run_dir` param to the workflow so that samples'
output will be sorted by run.