Fixed issue with pervious change to panel parsing
With the shift to the Genomics England gene panels, we changed the
way we are parsing the panels, pulling from the first column, instead
of just paring each line. The caused an extra pipe character to be
appended to the end of the matching expression. This change removes
that trailing pipe character.