Fixed some filtering
Fixed the CNV contig generation and applitcation of ClassifyCNV to
more accurately asses pathogenic CNVs. Also added the beginnings of a
JSON output file. The JSON output still needs to contain the genes
from the panel used for filtering, a list of steps to generate the
data, additional classifications for hits (and sort them accordingly).

Also need to add a classification field to the SNPs and SVs. will
look into adding variant classification to the expansions.