Update ufl-germline.nf
Commented out the `MANTA_WGS` process because it was causing issues in the pipeline run. This is still something I am looking into. It looks like there is an issue either downloading the bam file or allocating memory. This process works correctly when running independently so I am going to have to do some trial and error.