Update nextflow 20.10.0 > 21.10.6
Updated from nextflow 20.10.0 to 21.10.6 because it is far more up to date and fixes an issue we were having when running Manta. It looks like this issue was happening because Manta was run independently earlier and the job definition that existed from that was creating a conflict with the pipeline. It looks like this issue was corrected in 21.10.0, but we might as well be on the most up to date stable release.