Updated config
Updated nextflow.config to increase the amount of time inbetween download attempts and to try 8 times instead of 4 times. We have been seeing some process failures lately that are caused by lack of hard drive space. So we need to give more time for the EBS expanding storage to adjust up in size before downloading. Also updated the TODO to reflect this.