Update reporting.py
Changed the default gnomAD frequency to be -1 instead of 1. Looking back at the data, some alleles are present in the gnomAD exomes vcfs that are not present in the genomes vcf. This doesn't mean they are not real, this just means that we have to be more careful with the 2.1.1 data we are using.